test(assertions): Assertion tests were consolidated
