+++ vitest.config.ts
@@ -7,6 +7,9 @@ import basic from './vite.config.basic'
 
 export default mergeConfig(basic, defineConfig({
   test: {
+    include: [
+      '**/*.{test,spec}.?(c|m)[jt]s?(x)',
+    ],
     coverage: {
       provider: 'istanbul',
       include: ['src/**'],
